Updated model

+++ src/Svg.Model/Path/Commands/AddCirclePathCommand.cs
@@ -2,8 +2,8 @@
 {
     public class AddCirclePathCommand : PathCommand
     {
-        public double X { get; set; }
-        public double Y { get; set; }
-        public double Radius { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
+        public float Radius { get; set; }
     }
 }
