+++ src/Svg.Model/Path/Commands/AddRoundRectPathCommand.cs
@@ -3,7 +3,7 @@
     public class AddRoundRectPathCommand : PathCommand
     {
         public Rect Rect { get; set; }
-        public double RX { get; set; }
-        public double RY { get; set; }
+        public float RX { get; set; }
+        public float RY { get; set; }
     }
 }
