+++ src/Svg.Model/Path/Commands/CubicToPathCommand.cs
@@ -2,11 +2,11 @@
 {
     public class CubicToPathCommand : PathCommand
     {
-        public double X0 { get; set; }
-        public double Y0 { get; set; }
-        public double X1 { get; set; }
-        public double Y1 { get; set; }
-        public double X2 { get; set; }
-        public double Y2 { get; set; }
+        public float X0 { get; set; }
+        public float Y0 { get; set; }
+        public float X1 { get; set; }
+        public float Y1 { get; set; }
+        public float X2 { get; set; }
+        public float Y2 { get; set; }
     }
 }
