+++ src/Svg.Model/Path/Commands/LineToPathCommand.cs
@@ -2,7 +2,7 @@
 {
     public class LineToPathCommand : PathCommand
     {
-        public double X { get; set; }
-        public double Y { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
     }
 }
